Refactoring the code layout
Since the previous way of writing the code didn't work with the design that we chose, I have rewritten the code layout so it looks more like the actual code. No new functionality but I have added two more functions which don't do what I need it to but I'm still getting my head around how to add a quantity for a single product and not the full amount that they are ordering. If you look at line 142, it might give you a better idea of what I mean.